API Importer - Minor code change

+++ src/main/java/WoWSSSC/model/shipprofile/profile/Concealment.java
@@ -12,5 +12,5 @@ public class Concealment
     private float detect_distance_by_ship;
     private long total;
 
-    private double visibilityCoefGK;
+    private float visibilityCoefGK;
 }
